Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-audioworklet-interface/audioworkletnode-onerror.https-expected.txt

CONSOLE MESSAGE: line 50: TypeError: undefined is not an object (evaluating 'context.audioWorklet.addModule')

FAIL 
      Test onprocessorerror handler in AudioWorkletNode
     TypeError: undefined is not an object (evaluating 'context.audioWorklet.addModule')
